refactor(worker-sessions): consolidate control helpers
